Improved Camera Movement
I changed the rotation so the camera rotates about the y axis instead of loosy goosy however it wants. However, that created  another problem, where when you tried to pan around facing annything but forwards, you would move as if you were facing forwards. To fix this, I updated the movement logic so that for example, when you move the mouse right, the camera moves to it's local right, instead of the global "right". Thank you for coming to my TED talk.