Finished Dome Script
I made finished the updateInfo function in Dome.cs to correctly update names in the dome panel. For now, when it subtracts names due to the population lowering, it just chops the names off the end of the list, which I think is fine. I have not yet configured it in unity with the text, as I do not know how to do that with the prefabs.