More terrain shaping
A problem I'm having is making each terrian peice to need to go to height 0 at the edges so they line up, which looks a little funky but it's not too bad. I'm doing 3 layered approach: the tiles right around the middle that are realistic and low, higher mountains in the second layer, and really tall mountains the third layer.