Rover backend
I cooked up a rover manager script and got it all configured in unity so that you can view the stats of individual rovers, and it is all stored in the variable tracker. Now I just need to add so you can edit the values (buy upgrades for the rovers).