Made tutorial lowkey work on some nonchalant type shit
Added extra lines for gerald, and configured all movement.

Also, added a button in the pause menu to replay the tutorial.

Finally, made it remember if you've seen the tutorial or not, and do or don't play it accordingly.

TODO:
Add "i" icons to all buildings in shop that open panel that explain details about the panel.